# KaryawanApp.BlazorHybrid/Entity/Karyawan.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace KaryawanApp.BlazorHybrid.Entity
{
    public class Karyawan
    {
        public string? NIK { get; set; }
        public string? Nama { get; set; }
        public string? Divisi { get; set; }
        public string? Jabatan { get; set; }
        public string? Telp { get; set; }
        public string? Alamat { get; set; }
    }
}