# swaggerkaryawan.api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Karyawan API",
    "description": "API sederhana untuk CRUD data Karyawan",
    "version": "v1"
  },
  "paths": {
    "/api/Karyawan/create": {
      "post": {
        "tags": [
          "Karyawan"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Karyawan"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Karyawan"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Karyawan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Karyawan/update/{nik}": {
      "put": {
        "tags": [
          "Karyawan"
        ],
        "parameters": [
          {
            "name": "nik",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Karyawan"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Karyawan"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Karyawan"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Karyawan/get/{nik}": {
      "get": {
        "tags": [
          "Karyawan"
        ],
        "parameters": [
          {
            "name": "nik",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Karyawan/getall": {
      "get": {
        "tags": [
          "Karyawan"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Karyawan/delete/{nik}": {
      "delete": {
        "tags": [
          "Karyawan"
        ],
        "parameters": [
          {
            "name": "nik",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/": {
      "get": {
        "tags": [
          "KaryawanApi"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Karyawan": {
        "type": "object",
        "properties": {
          "nik": {
            "type": "string",
            "nullable": true
          },
          "nama": {
            "type": "string",
            "nullable": true
          },
          "jabatan": {
            "type": "string",
            "nullable": true
          },
          "telp": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}